+++ packages/component/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @orderly.network/components
 
+## 0.0.61
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.57
+  - @orderly.network/types@0.0.60
+  - @orderly.network/utils@0.0.48
+
 ## 0.0.60
 
 ### Patch Changes