+++ packages/tsconfig/CHANGELOG.md
@@ -1,5 +1,11 @@
 # tsconfig
 
+## 0.0.55
+
+### Patch Changes
+
+- update
+
 ## 0.0.54
 
 ### Patch Changes